Refactor & 1st part finished with mock
