+++ lib/core/di/service_locator.dart
@@ -35,12 +35,12 @@ import 'package:telegram/feature/auth/verify_mail/domain/repo/verify_mail_base_r
 import 'package:telegram/feature/auth/verify_mail/domain/use_case/send_otp_use_case.dart';
 import 'package:telegram/feature/auth/verify_mail/domain/use_case/verify_otp_use_case.dart';
 import 'package:telegram/feature/auth/verify_mail/presetnation/controller/verfiy_mail_cubit.dart';
-import 'package:telegram/feature/settings/Data/data_source/remote_data/user_settings_remote_data_source.dart';
-import 'package:telegram/feature/settings/Data/repositories/user_settings_repo_impl.dart';
-import 'package:telegram/feature/settings/Domain/repositories/user_settings_repo.dart';
-import 'package:telegram/feature/settings/Domain/use_cases/fetch_settings_use_case.dart';
-import 'package:telegram/feature/settings/Domain/use_cases/update_settings_use_case.dart';
-import 'package:telegram/feature/settings/Presentation/Controller/user_settings_cubit.dart';
+import 'package:telegram/feature/settings/datasettings/datasource/remotedata/user_settings_remote_data_source.dart';
+import 'package:telegram/feature/settings/datasettings/repos/user_settings_repo_impl.dart';
+import 'package:telegram/feature/settings/domainsettings/repos/user_settings_repo.dart';
+import 'package:telegram/feature/settings/domainsettings/usecases/fetch_settings_use_case.dart';
+import 'package:telegram/feature/settings/domainsettings/usecases/update_settings_use_case.dart';
+import 'package:telegram/feature/settings/presentationsettings/controller/user_settings_cubit.dart';
 import 'package:telegram/feature/splash_screen/presentation/controller/splash_cubit.dart';
 import 'package:telegram/feature/night_mode/presentation/controller/night_mode_cubit.dart';
