--- lib/feature/settings/Data/data_source/local_data/user_settings_local_data_source.dart
@@ -1,6 +1,0 @@
-import 'package:telegram/feature/settings/Data/model/user_settings_model.dart';
-
-abstract class UserSettingsLocalDataSource {
-  Future<UserSettingsBodyModel> fetchSettings();
-  Future<void> updateSettings(UserSettingsBodyModel settings);
-}
