+++ lib/feature/settings/domainsettings/entities/user_settings_entity.dart
