+++ lib/feature/settings/domainsettings/repos/user_settings_repo.dart
@@ -1,6 +1,6 @@
 import 'package:dartz/dartz.dart';
 import 'package:telegram/core/error/faliure.dart';
-import 'package:telegram/feature/settings/Domain/entities/user_settings_entity.dart';
+import 'package:telegram/feature/settings/domainsettings/entities/user_settings_entity.dart';
 
 abstract class UserSettingsRepo {
   Future<Either<Failure, UserSettingsEntity>> fetchSettings();
