+++ lib/feature/settings/domainsettings/usecases/fetch_settings_use_case.dart
@@ -1,8 +1,8 @@
 import 'package:dartz/dartz.dart';
 import 'package:telegram/core/error/faliure.dart';
 import 'package:telegram/core/use_cases/use_case.dart';
-import 'package:telegram/feature/settings/Domain/entities/user_settings_entity.dart';
-import 'package:telegram/feature/settings/Domain/repositories/user_settings_repo.dart';
+import 'package:telegram/feature/settings/domainsettings/entities/user_settings_entity.dart';
+import 'package:telegram/feature/settings/domainsettings/repos/user_settings_repo.dart';
 
 class FetchSettingsUseCase
     extends BaseUseCase<UserSettingsEntity, NoParameters> {
